requirements.txt
torch
numpy
pandas
matplotlib
seaborn
scikit-learn
PyYAML
sentence-transformers
faiss-cpu
beir
